Add "componentBlockStart" and "componentBlockEnd" to indexer
Ares needs access to the beginning and end of the components block -
i.e. everything between [ and ], in order to do proper source
replacement from the designer. The Indexer was the lowest-impact place
to add this. Ultimately, it'll be replaced by sending the AST directly
to the designer from the Parser, but I don't think I'll get that
running this week...

Also added a dead-simple components parsing test case to the
low-level-test.html file.