Fix accidental skipping of post-closure tokens in analyzer
ENYO-3484 showed a problem with parsing the code in enyo.Router.  The
root cause was code added for handling Enyo kinds defined in closures
that was too optimistic. It assumed the token after a closure that was
inside an association was the invocation, when this was also being
triggered for closures defined in calls to enyo.ready.  This resulted
in the following /**@public*/ pragma being eaten.

Also guarded two uses of node.future so code won't crash if you've
got a malformed source file.

Enyo-DCO-1.1-Signed-Off-By: Ben Combee (ben.combee@lge.com)